created color filter manager

+++ features/media/mediaeditor/src/main/java/com/tokopedia/media/editor/ui/component/BrightnessToolUiComponent.kt
@@ -20,8 +20,8 @@ class BrightnessToolUiComponent constructor(
 
         edtBrightnessValue.addTextChangedListener(object : TextWatcher {
             override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                val value = s.toString().toInt()
-                listener.onBrightnessValueChanged(if (value > 0) value else 0)
+                val value = s.toString().toFloat()
+                listener.onBrightnessValueChanged(if (value > 0f) value else 0f)
             }
 
             override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
@@ -31,7 +31,7 @@ class BrightnessToolUiComponent constructor(
     }
 
     interface Listener {
-        fun onBrightnessValueChanged(value: Int)
+        fun onBrightnessValueChanged(value: Float)
     }
 
 }
